I had my brother try out the keyboard and he said he had difficulty in discerning between each letter, suggesting we make vowels a different color. Thus each vowel is a different color. I then took it a bit further and made every other letter a different color.
But this leads me to my next point.

I don't know if we'll keep this change or modify it. 
I know we need to change the color scheme so it matches the default Xbox 360 keyboard and thus reduces bias in our later tests http://blog.gadgethelpline.com/wp-content/uploads/2010/03/XBOX-360-keyboard.bmp ). And if we match their color scheme we will be forced to remove the different colored keys as well.

Of course we /could/ still add this onto our analysis at the very least.